test: move WebSocket URL-instance proxy test to proxy.test.ts; revert websocket-proxy.test.ts
Touching websocket-proxy.test.ts at all (the earlier env-clearing and the new
subprocess test) correlated with the Docker-gated Squid proxy tests in that
file flaking on first run across several Linux lanes on every build, while
main shows zero such flakes. Reverting websocket-proxy.test.ts to main and
placing the WebSocket URL-instance test alongside the fetch URL-instance test
in proxy.test.ts (which already clears NO_PROXY file-level) keeps coverage
without perturbing the Squid tests.
